+++ rust-src/concordium_base/src/protocol_level_tokens/token_metadata_url.rs
@@ -146,7 +146,7 @@ mod tests {
             additional,
         };
 
-        let cbor_encoded = cbor_encode(&metadata_url).unwrap();
+        let cbor_encoded = cbor_encode(&metadata_url);
 
         let expected_cbor = Vec::from_hex("a4636b65796576616c75656375726c7368747470733a2f2f6578616d706c652e636f6d67616e6f7468657218286e636865636b73756d53686132353658200101010101010101010101010101010101010101010101010101010101010101").unwrap();
         let expected: MetadataUrl = cbor_decode(&expected_cbor).unwrap();